Import

+++ Pages/Transaction.cs
@@ -5,8 +5,8 @@
         public bool IsIncome { get; set; }
         public string Category { get; set; }
         public string Title { get; set; }
-        public int LastMonth { get; set; }
-        public int ThisMonth { get; set; }
+        public float LastMonth { get; set; }
+        public float ThisMonth { get; set; }
         public bool IsDone { get; set; }
     }
 }
